+++ CHANGELOG.md
@@ -7,7 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [TODO]
 
-- update trader restock to days remaining
-- update vend expiration to days remaining
+- update trader restock on day change
+- update trader restock on login
+- update trader restock ui to days remaining
 
-## [UNRELEASED]
+## [0.1.0] - 2023-04-14
+
+- add vend expiration buff
+- update vend expiration on day change
+- update vend expiration on login
+- update vend expiration on rent
+- update vend expiration ui to days remaining